+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.1.1 (2018-11-02)
+------------------
+
+* Fix related Pip files for requests module
+* Fixed test data generating function in merchant.py
+* Fixed tracking code for testing in test_get_shipping_label.py
+
 0.1.0 (2018-01-22)
 ------------------